backend/requirements.txt
# Core Dependencies
fastapi
uvicorn[standard]
pydantic
python-dotenv

# LangChain & LLM
langchain
langchain-groq
groq

# Google APIs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Database
aiosqlite

# HTTP & CORS
httpx
python-multipart


psycopg2-binary
sqlalchemy